pyproject.toml
[project]
name = "llm-warroom"
version = "0.1.0"
description = "Local-first AI reasoning workbench with Ask and War Room workflows"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.9.0",
    "openai>=2.0.0",
    "anthropic>=0.60.0",
]
